test(core): straight, flush, straigh_flush
add tests for straight, flush and straight flush detection. for straight flushes, I am not adding a test for every combination of 5 sequential cards(e.g. 2 to 6) and Suit, because the detection of same suit(i.e. flush) and straight(i.e. (2_to_6, 3_to_7, 4_to_8, 5_to_9) are tested independently. We are interested in these tests that a hand that is both straight and flush is detected properly.