ci(tests): ci tests
add ci pipeline and basic test step to aid with automating testing. generate and add requirements.txt. add utility script to generate requirements.txt using pipenv. pass tests whose scenarios are not yet implemented, because I only want to merge when the pipeline is green. the tests will be activated when the hand rank is implemented.